+++ src/variables/CookiesVariable.php
@@ -1,18 +1,16 @@
 <?php
 
 /**
- * Cookies plugin for Craft CMS 3.x
+ * Cookies plugin for Craft CMS
  *
  * @link      https://nystudio107.com/
- * @copyright Copyright (c) 2017 nystudio107
+ * @copyright Copyright (c) nystudio107
  * @license   MIT License https://opensource.org/licenses/MIT
  */
 
 namespace nystudio107\cookies\variables;
 
 use nystudio107\cookies\Cookies;
-
-use Craft;
 
 /**
  * Cookies template variables
@@ -23,17 +21,16 @@ use Craft;
  */
 class CookiesVariable
 {
-
     /**
      * Set a cookie
      *
      * @param string $name
      * @param string $value
-     * @param int    $expire
+     * @param int $expire
      * @param string $path
      * @param string $domain
-     * @param bool   $secure
-     * @param bool   $httpOnly
+     * @param bool $secure
+     * @param bool $httpOnly
      * @param string $sameSite
      */
     public function set(
@@ -44,7 +41,7 @@ class CookiesVariable
         $domain = "",
         $secure = false,
         $httpOnly = false,
-        $sameSite = null
+        $sameSite = 'Lax'
     ) {
         Cookies::$plugin->cookies->set(
             $name,
@@ -75,11 +72,11 @@ class CookiesVariable
      *
      * @param string $name
      * @param string $value
-     * @param int    $expire
+     * @param int $expire
      * @param string $path
      * @param string $domain
-     * @param bool   $secure
-     * @param bool   $httpOnly
+     * @param bool $secure
+     * @param bool $httpOnly
      * @param string $sameSite
      */
     public function setSecure(
@@ -90,7 +87,7 @@ class CookiesVariable
         $domain = "",
         $secure = false,
         $httpOnly = false,
-        $sameSite = null
+        $sameSite = 'Lax'
     ) {
         Cookies::$plugin->cookies->setSecure(
             $name,